Se borran productos individuales y otras cositas más que se van marcando en el documento 7/11/19

+++ app/Http/Controllers/CompanyController.php
@@ -680,9 +680,14 @@ class CompanyController extends Controller
             $history->description = "Se dio de baja el producto";
             $history->company = $id;
             $history->branch = $branch;
-            $history->save();
+            
+            
             $ser = License::where("serialkey",$qu1->serialkey)->first();
-            $ser->delete();
+            if($ser!=null){
+                $history->save();                
+                $ser->delete();
+            }
+            
         }
         
         return redirect()->route('showBranchesProducts',compact('id','branch'));
